imx6ul-grisp2.dts: move usb supplies out of phy
The USB power supplies were not being turned on. This resulted in the
WiFi module not getting power and therefore not showing up when
enumerating the USB bus.

The following warnings were in the kernel boot log:

```
[    0.892054] imx_usb 2184000.usb: 2184000.usb supply vbus not found, using dummy regulator
[    0.991934] imx_usb 2184200.usb: 2184200.usb supply vbus not found, using dummy regulator
```

Looking at the phy bindings [1], there's not a device tree binding for
`vbus-supply` in the phy.  However there's a `vbus-supply` in the USB
bindings [2], so move the `vbus-supply` there.

[1] https://elixir.bootlin.com/linux/v4.14.268/source/Documentation/devicetree/bindings/phy/mxs-usb-phy.txt
[2] https://elixir.bootlin.com/linux/v4.14.268/source/Documentation/devicetree/bindings/usb/ci-hdrc-usb2.txt#L34